% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicfuns.R
\name{beamSearchWrap}
\alias{beamSearchWrap}
\title{Wrapper function for beamsSearch}
\usage{
beamSearchWrap(pvec, y, onset, Npart, w0 = c(1, 0, 0, 0, 0, 0, 0, 0),
  npaths = 1)
}
\arguments{
\item{pvec}{a vector of the parameters for yupengMats in the order \eqn{\sigma_\eps^2, }}

\item{y}{the performed tempo of each note}

\item{onset}{a vector giving the measure number (including fracitons for notes within a measure) of the start of each note}

\item{Npart}{the desired number of particles}

\item{w0}{the prior probability of starting in each state}

\item{npaths}{the number of most likely paths to return}
}
\value{
list (length npaths) of lists. Each list contains the following: \describe{
\item{paths}{a vector giving the state path}
\item{weight}{the final sampling weight of the path}
\item{ests}{the smoothed tempo estimates from the path}
\item{llik}{the negative log liklihood for the path}
}
}
\description{
Wrapper function for beamsSearch
}
