% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicfuns.R
\name{convert10to4}
\alias{convert10to4}
\title{Converts from the expanded ten-state notation to four-state notation}
\usage{
convert10to4(path)
}
\arguments{
\item{path}{path in ten-state form}
}
\value{
path in four-state form
}
\description{
Because our model is actually 2-Markov, we expand from 4 to 16 states. 
But not all states can be visited, so there are only 10.
}
\examples{
path = sample(1:10, 25, replace=TRUE)
npath = convert10to4(path)

}
