% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicfuns.R
\name{convert4to2dynamics}
\alias{convert4to2dynamics}
\title{Converts from the expanded 4-state notation to 2-state notation}
\usage{
convert4to2dynamics(path)
}
\arguments{
\item{path}{path in four-state form}
}
\value{
path in two-state form
}
\description{
Because our model is actually 2-Markov, we expand from 2 to 4 states. 
All states can be visited.
}
\examples{
path = sample(1:10, 25, replace=TRUE)
npath = convert10to4(path)

}
