% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicfuns.R
\name{convert4to8}
\alias{convert4to8}
\title{Converts a 4 state 2-Markov path HMM to an 8 state 1-Markov path}
\usage{
convert4to8(path)
}
\arguments{
\item{path}{a vector of length n containing the integers 1 to 4}
}
\value{
a vector of length n containing the integers 0 to 7 (appropriate for C++)
}
\description{
This function is specific to the HMM given in Gu (2017)
}
