% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicfuns.R
\name{convert8to4}
\alias{convert8to4}
\title{Converts an 8 state 1-Markov path HMM to a 4 state 2-Markov path}
\usage{
convert8to4(path)
}
\arguments{
\item{path}{a vector of length n containing the integers 0 to 7 (as output by BeamSearch when applied to the yupengMats parameterization)}
}
\value{
a vector of length n containing the integers 1 to 4
}
\description{
This function is specific to the HMM given in Gu (2017)
}
