% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicfuns.R
\name{convertdynamicstates}
\alias{convertdynamicstates}
\title{Converts from the 0,1,2,3 to 1,2,3,4 notation}
\usage{
convertdynamicstates(path)
}
\arguments{
\item{path}{path in four-state form}
}
\value{
path in two-state form
}
\description{
Converts from the 0,1,2,3 to 1,2,3,4 notation
}
\examples{
path = sample(1:10, 25, replace=TRUE)
npath = convert10to4(path)

}
