% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdivergences.R
\name{Dist}
\alias{Dist}
\title{create a distance matrix}
\usage{
Dist(x, dist_fun = H2total, returnmat = TRUE)
}
\arguments{
\item{x}{data matrix}

\item{dist_fun}{distance function}

\item{returnmat}{boolean indicating whether to ensure that the return type is a matrix}
}
\value{
a distance matrix
}
\description{
create a distance matrix
}
