\name{dpf-package}
\alias{dpf-package}
\alias{dpf}
\docType{package}
\title{
\packageTitle{dpf}
}
\description{
\packageDescription{dpf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dpf}
\packageIndices{dpf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dpf}

Maintainer: \packageMaintainer{dpf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
