% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpf.R
\docType{package}
\name{dpf}
\alias{dpf}
\alias{dpf-package}
\title{dpf: Discrete particle filtering}
\description{
The dpf package provides three categories of important functions:
Kalman filters, a greedy discrete particle filter (beam search), and 
functions for applying these to musical tempo analysis
}
\section{Kalman filters}{
NA
}

\section{Beam search}{
NA
}

\section{Music tempo analysis}{
NA
}

