% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdivergences.R
\name{H2multinom}
\alias{H2multinom}
\title{Compute the squared Hellinger distance between two multinomial distributions}
\usage{
H2multinom(p, q, subtract1 = FALSE)
}
\arguments{
\item{p}{the probability vector for the first distribution}

\item{q}{the probability vector for the second distribution}

\item{subtract1}{a boolean indicating whether the squared distance should be computed instead of one minus the squared distance}
}
\value{
the squared Hellinger distance
}
\description{
Compute the squared Hellinger distance between two multinomial distributions
}
