% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdivergences.R
\name{H2total}
\alias{H2total}
\title{Compute the squared Hellinger distance between two models specified by passing the parameters to musicModel}
\usage{
H2total(p1, p2)
}
\arguments{
\item{p1}{parameters for the first model}

\item{p2}{parameters for the second model}
}
\value{
the squared Hellinger distance
}
\description{
Compute the squared Hellinger distance between two models specified by passing the parameters to musicModel
}
