% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdivergences.R
\name{hellinger_kde}
\alias{hellinger_kde}
\title{Estimate the Hellinger divergence between two distributions}
\usage{
hellinger_kde(x, y, index.list = 1, lb = 1e-06, kill = 1e-06,
  all = FALSE)
}
\arguments{
\item{x}{a sample from the first distribution (n x d)}

\item{y}{a sample from the second distribution (n x d)}

\item{index.list}{list of columns to be used (default 1), see details}

\item{lb}{lower-bound for the density truncation}

\item{all}{return the kde for x and y rather than the divergence estimate}
}
\value{
the Hellinger divergence
}
\description{
Estimate the Hellinger divergence between two distributions
}
\details{
This estimates the Hellinger divergence between samples from two 
distributions. The methodology implements the estimator in 
"Influence Functions for Machine Learning: Nonparametric Estimators for 
Entropies, Divergences and Mutual Informations" by Kandasamy, Krishnamurthy,
Poczos, Wasserman, and Robins. See their Matlab functions. Here, we use only
the Hellinger divergence.

Each sample (x and y) is assumed to be a matrix with n samples in d columns.
The estimator can only use up to 6 dimensions jointly. The list breaks up the
d dimensions into smaller collections. Generically,

\deqn{H = 2 - \int \sqrt(p_x p_y).}

If \eqn{p_x = p_{x1} p_{x2}} and \eqn{p_y = p_{y1}p_{y2}}, then 
\deqn{H = 2 - \int \sqrt(p_{x1} p_{y1}) \int \sqrt(p_{x1} p_{y1}).} So, for example, to 
treat the first column as independent from columns 2 and 3, use
\code{index.list = list(1,2:3)}.
}
\examples{
x = matrix(rnorm(3*1000)) \%*\% matrix(c(1,0,0,0,2,.5,0,.5,4),ncol=3)
y = matrix(rnorm(3*1000)) \%*\% matrix(c(2,0,0,0,1,.5,0,.5,4),ncol=3)
hellinger_kde(x, y, list(1, 2:3))
}
