% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kalman}
\alias{kalman}
\title{Estimate continuous states given parameters and discrete hidden states}
\usage{
kalman(pmats, path, y)
}
\arguments{
\item{pmats}{a list with componente \code{a0, P0, dt, ct, Tt, Zt, HHt, and GGt} 
e.g., as output from musicModel. See the details below.}

\item{path}{path of discrete hidden states}

\item{y}{observations}
}
\value{
List with components from Kalman filter and Smoother. See details.
}
\description{
Performs Kalman filtering and smoothing for the switching state space model.
Also works on any (potentially time varying) state space model.
}
\examples{
data(tempos)
theta = c(363.475,125.05,-11.81757,-34.91605,366.5073,0.519518,
          0.212499,0.7901238,0.0989846,0.7815691,0.2194727,0.03586685,
          0.1342895,0.2435351)
y = matrix(tempos[,'Milkina_1970'], 1)
lt = diff(c(tempos$note_onset, 61))
pmats = musicModel(lt, theta[1], theta[2:4], theta[5:7], theta[8:14], 
                  c(132,0), c(400,10)) # prior means and variances on X_1
beam = with(pmats, beamSearch(a0, P0, c(1,0,0,0,0,0,0,0,0,0), dt, ct, Tt, Zt,
            HHt, GGt, y, transMat, 200))
bestpath = with(beam, paths[which.max(weights),])
kal = kalman(pmats, bestpath, y)

}
