% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{musicModel}
\alias{musicModel}
\title{Parameter matrices for our music model}
\usage{
musicModel(lt, sig2eps, mus, sig2eta, transprobs, initialMean,
  initialVariance)
}
\arguments{
\item{lt}{vector of durations between successive notes in the score}

\item{sig2eps}{variance of the observation noise}

\item{mus}{vector of 3 mean parameters (\eqn{\mu, \tau, and \varphi})}

\item{sig2eta}{vector of 3 state variance parameters (\eqn{\sigma_3^2, \sigma_2^2,and \sigma_4^2})}

\item{transprobs}{vector of 7 transition probabilities}

\item{initialMean}{a vector of length 2 giving the prior tempo and the prior acceleration for when state 1 or 3 is the starting state}

\item{initialVariance}{a vector of length 2 giving the prior variance for the tempo and the prior variance for the acceleration for when state 1 or 3 is the starting state}
}
\value{
List with components as appropriate for Kalman filtering or Beam Search. These include: \describe{
\item{a0}{a pxd matrix of the initial means of the hidden state. The j'th column corresponds to the initial mean when starting in the j'th discrete state.}
\item{P0}{a (p^2)xd matrix of the initial covariances of the hidden state. The j'th column corresponds to the initial covariances stored columnwise when starting in the j'th discrete state.}
\item{dt}{a pxdxn cube of state intercepts. The j'th column of the i'th slice corresponds to the intercept specified by the j'th discrete state at time i.}
\item{ct}{a kxdx1 cube of observation intercepts. The j'th column corresponds to the intercept specified by the j'th discrete state.}
\item{Tt}{a (p^2)xdxn cube of state slopes. The j'th column of the i'th slice corresponds to the slope matrix stored columnwise of the j'th discrete state at time i.}
\item{Zt}{a pkxdx1 cube of obvervation slopes. The j'th column corresponds to the slope matrix stored columnwise of the j'th discrete state.}
\item{HHt}{a (p^2)xdxn cube of state covariances. The j'th column of the i'th slice corresponds to the covariance matrix stored columnwise of the j'th discrete state at time i.}
\item{GGt}{a (k^2)xdx1 cube of observation covariances. The j'th column corresponds to the covariance matrix stored columnwise of the j'th discrete state.}
\item{transMat}{a dxd matrix of transition probabilities for the discrete states}
}
}
\description{
This function accepts a number of parameters and creates a list of matrices
for Kalman filter evaluation. See the paper for the particular form of the model.
}
\examples{
data(tempos)
theta = c(363.475,125.05,-11.81757,-34.91605,366.5073,0.519518,
          0.212499,0.7901238,0.0989846,0.7815691,0.2194727,0.03586685,
          0.1342895,0.2435351)
y = matrix(tempos[,'Milkina_1970'], 1)
lt = diff(c(tempos$note_onset, 61))
pmats = musicModel(lt, theta[1], theta[2:4], theta[5:7], theta[8:14], 
                  c(132,0), c(400,10))
                  
}
