% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStates}
\alias{plotStates}
\title{Plots the observed tempo and estimated discrete states given parameters.}
\usage{
plotStates(performance, params, y, onset, particleNumber = 200,
  initialMean = c(132, 0), initialVariance = c(400, 10))
}
\arguments{
\item{performance}{The named performance, e.g. 'Richter_1976'. See \code{names(tempos)}.}

\item{params}{A vector of parameters of length 14. No checks are performed.}

\item{y}{A vector of tempos.}

\item{onset}{A vector of note onset times.}

\item{particleNumber}{Number of particles for \code{beamSearch()}. Default is 200.}

\item{initialMean}{Mean for the first note in constant tempo. Length 2. Default is (132,0).}

\item{initialVariance}{Variance for the first note in constant tempo. Length 2. Default is (400,0).}
}
\description{
Plots the observed tempo and estimated discrete states given parameters.
}
\examples{
params = c(426.69980736, 136.33213703, -11.84256691, -34.82234559, 
          439.37886221, 1, 1, 0.84916635, 0.04611644, 0.74119571, 
          0.43966082, 0.02116317, 0.24513563, 0.17253254)
data(tempos)
y = tempos[,'Richter_1976']
onset = tempos$note_onset
plotStates('Richter_1976', params, y, onset)
}
