% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStates2}
\alias{plotStates2}
\title{Plots the Music Model states and Kalman smoothed estimates}
\usage{
plotStates2(
  theta,
  states,
  y,
  onset,
  model,
  priormean,
  priorvar,
  title = "Music Plot"
)
}
\arguments{
\item{theta}{parameters in same order}

\item{states}{vector of states at each time period (not converted)}

\item{y}{vector of tempos or dynamics of a single performance}

\item{onset}{vector of onsets.}

\item{model}{If using the dynamics model put "dynamics".  If using the tempo model put "tempo".}

\item{priormean}{Prior Mean array}

\item{priorvar}{Prior Variance array}

\item{title}{Change the title of the plot}
}
\value{
A plot showing the estimated values of the kalman smoother along with the states.
}
\description{
Plots the Music Model states and Kalman smoothed estimates
}
\examples{
theta = c(426.69980736, 136.33213703, -11.84256691, -34.82234559, 
         439.37886221, 1, 1, 0.84916635, 0.04611644, 0.74119571, 
         0.43966082, 0.02116317, 0.24513563, 0.17253254)
y = matrix(tempos[,'Richter_1976'], 1)
lt = diff(c(tempos$note_onset, 61))
pmats = musicModel(lt, theta[1], theta[2:4], theta[5:7], theta[8:14], 
                  c(132,0), c(400,10)) # prior means and variances on X_1
beam = with(pmats, beamSearch(a0, P0, c(1,0,0,0,0,0,0,0,0,0), dt, ct, Tt, Zt,
                             HHt, GGt, y, transMat, 200))
states = beam$paths[which.max(beam$weights),]

plotStates2(theta,states,y,tempos$note_onset,model="tempo",c(132,0),c(400,10),title="Richter 1976")


theta = c(20,10,
         12,13,14,
         21,22,23,
         .3,.2,.1,.5)
y = matrix(dynamics[,'Richter_1976'], 1)
lt = diff(c(dynamics$note_onset, 61))

pmats = musicModeldynamics(lt, theta[1], theta[2], theta[3:5], theta[6:8], theta[9:12], c(20,0,0), c(25,10,0)) 
beam = with(pmats, beamSearch(a0, P0, c(1,0,0,0), dt, ct, Tt, Zt,
                             HHt, GGt, y, transMat, 200))
states = beam$paths[which.max(beam$weights),]
kal <- kalman(pmats,states,y)
plotStates2(theta,states,y,dynamics$note_onset,model="dynamics",c(20,0,0), c(25,10,0),title="Richter 1976")

}
