% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\alias{ddirichlet}
\title{Dirichlet random variables}
\usage{
rdirichlet(n, K = length(alpha), alpha = rep(1, K))

ddirichlet(x, alpha, log = TRUE, normalize = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{K}{number of components}

\item{alpha}{vector of parameters. Must be positive. These functions are not vectorized over alpha.}

\item{x}{vector of probabilities supported on the K-1 simplex. ddirichlet is vectorized over x.}
}
\value{
\code{ddirichlet} gives the density on the log scale, unnormalized by default. \code{rdirichlet} generates random variates.
}
\description{
Dirichlet random variables
}
\examples{

randos = rdirichlet(10, 3, c(7,2,1))
density = ddirichlet(c(.5,.5,.5), c(7,2,1))
}
