% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpf.R
\docType{data}
\name{recordings}
\alias{recordings}
\title{Recordings}
\format{A data frame with 46 rows and 2 variables:
\describe{
  \item{performer}{The last name of the pianist}
  \item{year}{The year in which the recording was made}
}}
\source{
\url{http://www.mazurka.org.uk}
}
\usage{
recordings
}
\description{
A dataset containing the performers and years of performance of 46 recordings of Chopin's Mazurka Op. 68, No. 3
}
\keyword{datasets}
