% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{yupengMats}
\alias{yupengMats}
\title{Create parameter matrices as in Gu (2017)}
\usage{
yupengMats(lt, sig2eps, mus, sig2eta, transprobs, initialMean,
  initialVariance)
}
\arguments{
\item{lt}{durations between successive notes in the score}

\item{sig2eps}{variance of the observation noise}

\item{mus}{vector of 3 mean parameters}

\item{sig2eta}{vector of 3 state variance parameters}

\item{transprobs}{vector of 4 transition probabilities}
}
\value{
List with components as appropriate for Kalman filtering or Beam Search
}
\description{
Create parameter matrices as in Gu (2017)
}
