% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beamSearch}
\alias{beamSearch}
\title{Greedy HMM estimation given continuous hidden states}
\usage{
beamSearch(a0, P0, w0, dt, ct, Tt, Zt, HHt, GGt, yt, transProbs, N)
}
\arguments{
\item{a0}{a px1 matrix of state prior means}

\item{P0}{a pxp matrix of state}

\item{dt}{dx1 or dxn matrix of}
}
\value{
List with components "paths", "weights", and "LastStep"
}
\description{
Greedy HMM estimation given continuous hidden states
}
