% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicfuns.R
\name{convert10to4}
\alias{convert10to4}
\title{convert from the expanded ten-state notation to four-state notation}
\usage{
convert10to4(path)
}
\arguments{
\item{path}{path in ten-state form}
}
\value{
path in four-state form
}
\description{
convert from the expanded ten-state notation to four-state notation
}
