% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/dpf.R
\docType{package}
\name{dpf}
\alias{dpf}
\alias{dpf-package}
\title{Move the particles forward in time one step}
\usage{
dpf(currentStates, w, N, transProbs, a0, P0, dt, ct, Tt, Zt, HHt, GGt, yt)
}
\arguments{
\item{currentStates}{a vector of the current discrete state for each particle}

\item{w}{a vector of the sampling weights for each particle}

\item{N}{the maximum particle number}

\item{transProbs}{a dxd matrix of transition probabilities for the discrete states}

\item{a0}{a pxN matrix of the current estimate of the state means. Each column represents a particle}

\item{P0}{a (p^2)xN state prior covariance matrix}

\item{dt}{a pxd matrix of state intercepts. The j'th column corresponds to the intercept specified by the j'th discrete state.}

\item{ct}{a kxd matrix of observation intercepts. The j'th column corresponds to the intercept specified by the j'th discrete state.}

\item{Tt}{a (p^2)xd matrix of state slopes. The j'th column corresponds to the slope matrix stored columnwise of the j'th discrete state.}

\item{Zt}{a pkxd matrix of obvervation slopes. The j'th column corresponds to the slope matrix stored columnwise of the j'th discrete state.}

\item{HHt}{a (p^2)xd matrix of state covariances. The j'th column corresponds to the covariance matrix stored columnwise of the j'th discrete state.}

\item{GGt}{a (k^2)xd matrix of observation covariances. The j'th column corresponds to the covariance matrix stored columnwise of the j'th discrete state.}

\item{yt}{a kxn matrix of obervations}
}
\description{
The dpf package provides three categories of important functions:
Kalman filters, a greedy discrete particle filter (beam search), and 
functions for applying these to musical tempo analysis.
}
\details{
For a concise description, see the package vignette.

The most recent version of this package is available on github, and installable
with

\code{devtools::install_github('dajmcdon/dpf', build_vignettes=TRUE)}
}
