% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpf.R
\docType{package}
\name{dpf}
\alias{dpf}
\title{dpf: Discrete particle filtering}
\description{
The dpf package provides three categories of important functions:
Kalman filters, a greedy discrete particle filter (beam search), and 
functions for applying these to musical tempo analysis.
}
\details{
For a concise description, see the package vignette.

The most recent version of this package is available on github, and installable
with

\code{devtools::install_github('dajmcdon/dpf', build_vignettes=TRUE)}
}
