% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getloglike}
\alias{getloglike}
\title{Evaluate the likelihood given parameters and discrete states}
\usage{
getloglike(pmats, path, y)
}
\arguments{
\item{pmats}{a list of parameter matrices for the kalman filter. This can either be the output of yupengMats, or a list with the same names as the output of yupengMats.}

\item{path}{vector giving the desired path for hidden discrete states}

\item{y}{observations, each time point in a column}
}
\value{
the negative log-likelihood
}
\description{
Evaluate the likelihood given parameters and discrete states
}
