% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdivergences.R
\name{H2gauss}
\alias{H2gauss}
\title{Title}
\usage{
H2gauss(m1, m2, s1, s2, subtract1 = FALSE)
}
\arguments{
\item{m1}{}

\item{m2}{}

\item{s1}{}

\item{s2}{}

\item{subtract1}{}
}
\description{
Title
}
