% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kalman}
\alias{kalman}
\title{Estimate continuous states given parameters and discrete hidden states}
\usage{
kalman(pmats, path, y)
}
\arguments{
\item{pmats}{e.g., as output from musicModel}

\item{path}{path of discrete hidden states}

\item{y}{observations}
}
\value{
List with components from Kalman filter and Smoother
}
\description{
Estimate continuous states given parameters and discrete hidden states
}
