% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kalman}
\alias{kalman}
\title{Estimate continuous states given parameters and discrete hidden states}
\usage{
kalman(pmats, path, y)
}
\arguments{
\item{pmats}{a list with componente \code{a0, P0, dt, ct, Tt, Zt, HHt,} and \code{GGt} 
e.g., as output from musicModel. See the details below.}

\item{path}{path of discrete hidden states}

\item{y}{observations}
}
\value{
List with components from Kalman filter and Smoother. See details.
}
\description{
Performs Kalman filtering and smoothing for the switching state space model.
Also works on any (potentially time varying) state space model.
}
\details{

}
\examples{
data(tempos)
theta = c(426.69980736, 136.33213703, -11.84256691, -34.82234559, 
          439.37886221, 1, 1, 0.84916635, 0.04611644, 0.74119571, 
          0.43966082, 0.02116317, 0.24513563, 0.17253254)
y = matrix(tempos[,'Richter_1976'], 1)
lt = diff(c(tempos$note_onset, 61))
pmats = musicModel(lt, theta[1], theta[2:4], theta[5:7], theta[8:14], 
                  c(132,0), c(400,10)) # prior means and variances on X_1
beam = with(pmats, beamSearch(a0, P0, c(1,0,0,0,0,0,0,0,0,0), dt, ct, Tt, Zt,
            HHt, GGt, y, transMat, 200))
bestpath = with(beam, paths[which.max(weights),])
kal = kalman(pmats, bestpath, y)

}
