% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStates}
\alias{plotStates}
\title{Plot the state path found by beamSearch along with the smoothed tempo estimate}
\usage{
plotStates(performance, performer, onset, params, particleNumber = 200,
  initialMean = c(132, 0), initialVariance = c(400, 10))
}
\arguments{
\item{performance}{a vector of the performed tempo of each note}

\item{performer}{a string giving the name of the performer}

\item{onset}{a vector giving the measure number (including fracitons for notes within a measure) of the start of each note}

\item{particleNumber}{the desired number of particles to be used in beamSearch (default 200)}

\item{initialVariance}{a vector to be passed to yupengMats (default is c(400,10))}

\item{initialMeans}{a vector to be passed to yupengMats (default is c(132,0))}
}
\description{
Plot the state path found by beamSearch along with the smoothed tempo estimate
}
