% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStates}
\alias{plotStates}
\title{Title}
\usage{
plotStates(performance, performer, onset, params, particleNumber = 200,
  initialMean = c(132, 0), initialVariance = c(400, 10))
}
\arguments{
\item{performance}{}

\item{performer}{}

\item{onset}{}

\item{params}{}

\item{particleNumber}{}

\item{initialMean}{}

\item{initialVariance}{}
}
\description{
Title
}
